Move @ariakit/test browser shim logic into shims.ts
Move applyBrowserPolyfills (renamed applyBrowserShims), the happy-dom
shim helpers, and the window.alert shim out of __utils.ts and into
shims.ts, so all the browser-shim logic lives with the module that
applies it at import. __utils.ts keeps only environment detection, the
flush/frame helpers, and wrapAsync (which now only toggles the act
environment). Behavior is unchanged.
